Revert "asoc: codecs: Update clk_div_get returned type"
This reverts commit b83d1dfee3f8f5f62a988aae74810c6e68a5c82b.

+++ asoc/codecs/lpass-cdc/lpass-cdc-tx-macro.c
@@ -1833,7 +1833,7 @@ static int lpass_cdc_tx_macro_clk_div_get(struct snd_soc_component *component)
 	if (!lpass_cdc_tx_macro_get_data(component, &tx_dev, &tx_priv, __func__))
 		return -EINVAL;
 
-	return (int)tx_priv->dmic_clk_div;
+	return tx_priv->dmic_clk_div;
 }
 
 static int lpass_cdc_tx_macro_validate_dmic_sample_rate(u32 dmic_sample_rate,
